Fix timezone offset for western timezones, add File.DataOffset()
The recording timestamp timezone offset is a signed byte (15-minute
intervals from GMT); reading it as unsigned produced wrong times for
western (negative-offset) timezones. Sign-extend it via int8.

Also adds File.DataOffset(), returning the byte offset of a file's
data within the ISO image.

Both changes adapted from PR #3 by @pschou (the go.mod/go.sum changes
in that PR, which retargeted the module path, were dropped).

+++ iso9660.go
@@ -558,7 +558,9 @@ func (ts *RecordingTimestamp) UnmarshalBinary(data []byte) error {
 	hour := int(data[3])
 	min := int(data[4])
 	sec := int(data[5])
-	tzOffset := int(data[6])
+	// the timezone offset is a signed byte (number of 15-minute intervals
+	// from GMT), so negative offsets (western timezones) must be sign-extended
+	tzOffset := int(int8(data[6]))
 	secondsInAQuarter := 60 * 15
 
 	tz := time.FixedZone("", tzOffset*secondsInAQuarter)
